requirements.txt
fastapi
uvicorn
openai
faiss-cpu
pypdf
tiktoken
requests
beautifulsoup4